cloud/gcp/thumbnail-generator/modules/function/playbooks/function/requirements.txt
google-cloud-storage==1.29.0
Wand==0.6.2
